Connection Type as Enum UDA & sending requires the connections.
Connection Type as Enum UDA: SO much more user friendly, if only D bug #20835 was not making it hacking to do.

If a connection type is given for a target, then connections must be given to the send function.

If multiple targets are given then connections must be given in the order of the enum.